+++ backend/src/main/java/com/net1707/backend/dto/CartItemDTO.java
@@ -5,13 +5,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class CartItemDTO {
-    private Integer productId;
+    private Long productId;
     private String productName;
     private Integer quantity;
-    private float price;
+    private BigDecimal price;
 }
