+++ backend/src/main/java/com/net1707/backend/dto/OrderDetailResponseDTO.java
@@ -2,6 +2,8 @@ package com.net1707.backend.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Data
@@ -9,8 +11,8 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderDetailResponseDTO {
-    private int productId;
-    private Integer batchId;
+    private Long productId;
+    private Long batchId;
     private Integer quantity;
-    private Float unitPrice;
+    private BigDecimal unitPrice;
 }
