+++ backend/src/main/java/com/net1707/backend/model/ProductBatch.java
@@ -20,7 +20,7 @@ public class ProductBatch {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer batchId;
+    private Long batchId;
 
     @ManyToOne
     @JoinColumn(name = "ProductID")
